documented in human readable form modules, renamed helperfunctions to utils and sdc to sdata, finished about section
