+++ data_generation/docs/source/acknowledgements.rst
@@ -1,52 +1,23 @@
-.. _about:
-
-About
-=====
-
-.. _contact:
-
-Contact
--------
-
-Questions can be directed to `Franziska Menti <https://fmenti.github.io/>`_ (fmenti@phys.ethz.ch) who has been the main developer during the recent years.
-
-.. _attribution:
-
-Attribution
------------
-
-If you use life_td or LIFE-StarCat version 4 or newer in your publication then pleace cite `Menti F.  et al: The LIFE Target Star Database LIFETD, TAP-published table set, doi:10.21938/ke_e6lzO_jjX_vzvVIcwZA <https://dc.zah.uni-heidelberg.de/voidoi/q/lp/custom/10.21938/ke_e6lzO_jjX_vzvVIcwZA>`_.
-
-.. _contribution:
-
-Contribution
-------------
-
-Join the `LIFE Target Database Work Group <https://life-space-mission.com/contact/>`_ by following the instructions on the LIFE webpage.
-
+.. _acknowledgements:
 
 Acknowledgements
-----------------
+================
 
-Funding
-^^^^^^^
-
-ETH Zurich
+We acknowledge the continuing fruitful and constructive cooperation between `Workaut <workaut.ch>`_ and ETH Zurich that makes this project possible.
 
 Data and Facilities
-^^^^^^^^^^^^^^^^^^^ 
+------------------- 
 
 * This research has made use of the Washington Double Star Catalog maintained at the U.S. Naval Observatory.
 * This work has made use of data from the European Space Agency (ESA) mission `Gaia <https://www.cosmos.esa.int/gaia>`_ ,processed by the Gaia Data Processing and Analysis Consortium `DPAC <https://www.cosmos.esa.int/web/gaia/dpac/consortium>`_. Funding for the DPAC has been provided by national institutions, in particular, the institutions participating in the Gaia Multilateral Agreement.
 * This research has made use of the `SIMBAD <http://simbad.u-strasbg.fr/simbad/>`_ database, operated at CDS, Strasbourg, France.
 * This research has made use of the `VizieR <https://vizier.cds.unistra.fr/viz-bin/VizieR>`_ catalog access tool, CDS, Strasbourg, France (DOI: 10.26093/cds/vizier).
-* grant kennedy
-* exomercat
+* This research has made use of the star database `sdb <http://drgmk.com/sdb/>`_ by Grant Kennedy which is a combination of published and unpublished data, plus SED fitting using the method in the Yelverton papers (`2019 <https://ui.adsabs.harvard.edu/abs/2019MNRAS.488.3588Y/abstract>`_ and `2020 <https://ui.adsabs.harvard.edu/abs/2020MNRAS.495.1943Y/abstract>`_).
+* This research has made use of `Exo-Mercat (Alei et al. 2020) <https://ui.adsabs.harvard.edu/abs/2020A%26C....3100370A/abstract>`_
 * This research has made use of the relation published in Table 5 of `Pecaut & Mamajek (2013) <https://ui.adsabs.harvard.edu/abs/2013ApJS..208....9P/abstract>`_.
 
-
 Software
-^^^^^^^^ 
+-------- 
 
 * This work made use of NumPy `(Harris et al. 2020) <https://www.nature.com/articles/s41586-020-2649-2>`_.
 * This work made use of `Astropy <http://www.astropy.org>`_ a community-developed core Python package and an ecosystem of tools and resources for astronomy `(2013) <https://ui.adsabs.harvard.edu/abs/2013A%26A...558A..33A/abstract>`_, `(2018) <https://ui.adsabs.harvard.edu/abs/2018AJ....156..123A/abstract>`_, `(2022) <https://ui.adsabs.harvard.edu/abs/2022ApJ...935..167A/abstract>`_. 
@@ -54,7 +25,7 @@ Software
 * This work made use of GAVO DaCHS `2014A&C…..7…27D <https://ui.adsabs.harvard.edu/abs/2014A%26C.....7...27D/abstract>`_.
 * This work made use of matplotlib `(Hunter 2007) <https://ieeexplore.ieee.org/document/4160265>`_.
 
-
 Other
-^^^^^
+-----
+
 This work made use of the book "Beginning Database Design: From Novice to Professional" `Churcher 2012 <https://books.google.ch/books?id=6ksf6yHS7L4C>`_
