+++ data_generation/docs/source/attribution.rst
@@ -1,0 +1,6 @@
+.. _attribution:
+
+Attribution
+===========
+
+If you use life_td or LIFE-StarCat version 4 or newer in your publication then pleace cite `Menti F.  et al: The LIFE Target Star Database LIFETD, TAP-published table set, doi:10.21938/ke_e6lzO_jjX_vzvVIcwZA <https://dc.zah.uni-heidelberg.de/voidoi/q/lp/custom/10.21938/ke_e6lzO_jjX_vzvVIcwZA>`_.
