+++ data_generation/docs/source/contact.rst
@@ -1,0 +1,7 @@
+.. _contact:
+
+Contact
+=======
+
+Questions can be directed to `Franziska Menti <https://fmenti.github.io/>`_ (fmenti@phys.ethz.ch) who has been the main developer during the recent years.
+
