+++ data_generation/docs/source/contribution.rst
@@ -1,0 +1,6 @@
+.. _contribution:
+
+Contribution
+============
+
+Join the `LIFE Target Database Work Group <https://life-space-mission.com/contact/>`_ by following the instructions on the LIFE webpage.
